Update path for new device agents to testflinger_device_connectors
