+++ src/testflinger_device_connectors/devices/lenovo_oemscript/lenovo_oemscript.py
@@ -12,17 +12,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Ubuntu OEM Script provisioning for Dell OEM devices
+"""Ubuntu OEM Script provisioning for Lenovo OEM devices
 this for systems that can use the oem recovery-from-iso.sh script
 for provisioning, but require the --ubr flag in order to use the
 "ubuntu recovery" method.
 """
 
 import logging
-from snappy_device_agents.devices.oemscript.oemscript import OemScript
+from testflinger_device_connectors.devices.oemscript.oemscript import OemScript
 
 logger = logging.getLogger()
 
 
-class DellOemScript(OemScript):
-    """Device Agent for Dell OEM devices."""
+class LenovoOemScript(OemScript):
+    """Device Agent for Lenovo OEM devices."""
+
+    # Extra arguments to pass to the OEM script
+    extra_script_args = ["--ubr"]
